+++ src/main/java/org/opensearch/neuralsearch/processor/TextEmbeddingProcessor.java
@@ -56,7 +56,49 @@ public final class TextEmbeddingProcessor extends InferenceProcessor {
         Environment environment,
         ClusterService clusterService
     ) {
-        super(tag, description, batchSize, TYPE, LIST_TYPE_NESTED_MAP_KEY, modelId, fieldMap, clientAccessor, environment, clusterService);
+        this(
+            tag,
+            description,
+            batchSize,
+            DEFAULT_BATCH_SIZE_BYTES,
+            modelId,
+            fieldMap,
+            skipExisting,
+            textEmbeddingInferenceFilter,
+            openSearchClient,
+            clientAccessor,
+            environment,
+            clusterService
+        );
+    }
+
+    public TextEmbeddingProcessor(
+        String tag,
+        String description,
+        int batchSize,
+        int batchSizeBytes,
+        String modelId,
+        Map<String, Object> fieldMap,
+        boolean skipExisting,
+        TextEmbeddingInferenceFilter textEmbeddingInferenceFilter,
+        OpenSearchClient openSearchClient,
+        MLCommonsClientAccessor clientAccessor,
+        Environment environment,
+        ClusterService clusterService
+    ) {
+        super(
+            tag,
+            description,
+            batchSize,
+            batchSizeBytes,
+            TYPE,
+            LIST_TYPE_NESTED_MAP_KEY,
+            modelId,
+            fieldMap,
+            clientAccessor,
+            environment,
+            clusterService
+        );
         this.skipExisting = skipExisting;
         this.textEmbeddingInferenceFilter = textEmbeddingInferenceFilter;
         this.openSearchClient = openSearchClient;